services

+++ app/Services/Service/ServService.php
@@ -4,11 +4,15 @@ namespace App\Services\Service;
 
 use App\Models\Service;
 use App\Http\Resources\ServiceResource;
+use App\Http\Resources\UserResource;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Requests\Api\Service\StoreServiceRequest;
 use App\Http\Requests\Api\Service\UpdateServiceRequest;
 use Illuminate\Support\Facades\Storage;
+use App\Constants\Constants;
+use App\Http\Requests\Api\Service\ChangeServiceStatusRequest;
+
 
 class ServService
 {
@@ -18,6 +22,10 @@ class ServService
     public function getAll($trashOnly = false, $paginate = false, $limit = 10)
     {
         $servicesQuery = Service::query();
+
+        if (auth()->user()->hasRole(Constants::SERVICE_PROVIDER_ROLE)) {
+            $servicesQuery = auth()->user()->services();
+        }
 
         // Apply soft delete filter if requested
         if ($trashOnly) {
@@ -29,13 +37,14 @@ class ServService
         // Load the associated category and provider relationships
         $servicesQuery->with(['category', 'provider', 'reviews']);
 
-        // Paginate the results if needed
-        if ($paginate) {
-            return $servicesQuery->paginate($limit);
-        }
+        $services = $paginate ? $servicesQuery->paginate($limit) : $servicesQuery->get();
 
-        return $servicesQuery->get();
-    }
+        // If the logged-in user is a normal user, return as ServiceResource
+        if (auth()->user()->hasRole(Constants::USER_ROLE)) {
+            return ServiceResource::collection($services);
+        }
+            return $servicesQuery->get();
+        }
 
     /**
      * Store a new service.
@@ -49,15 +58,27 @@ class ServService
             $data['image'] = $request->file('image')->storePublicly('services/images', 'public');
         }
 
+        if (auth()->user()->hasRole(Constants::ADMIN_ROLE)) {
+            if ($request->has('service_provider_id')) {
+                $data['service_provider_id'] = $request->input('service_provider_id');
+            } 
+        } elseif (auth()->user()->hasRole(Constants::SERVICE_PROVIDER_ROLE)) {
+            $data['service_provider_id'] = auth()->id();
+        } 
+
         // Store the service data
         $data['name'] = [
             'ar' => $data['ar_name'],
             'en' => $data['en_name'],
         ];
+        $data['description'] = [
+            'ar' => $data['ar_description'],
+            'en' => $data['en_description'],
+        ];
 
         $service = Service::create($data);
 
-        return new ServiceResource($service);
+        return $service;
     }
 
     /**
@@ -66,6 +87,10 @@ class ServService
     public function update(UpdateServiceRequest $request, Service $service)
     {
         $data = $request->validated();
+
+        if (auth()->user()->hasRole(Constants::SERVICE_PROVIDER_ROLE) && $service->service_provider_id !== auth()->id()) {
+            return error(__('messages.unauthorized_action'), [], 403);
+        }
 
         // Handle file upload if exists
         if ($request->hasFile('image')) {
@@ -76,6 +101,11 @@ class ServService
                 Storage::delete("public/$service->image");
             }
         }
+        if (auth()->user()->hasRole(Constants::ADMIN_ROLE)) {
+            if ($request->has('service_provider_id')) {
+                $data['service_provider_id'] = $request->input('service_provider_id');
+            } 
+        }
 
         // Update service name in different languages
         if (isset($data['ar_name'])) {
@@ -84,11 +114,17 @@ class ServService
         if (isset($data['en_name'])) {
             $data['name']['en'] = $data['en_name'];
         }
+        if (isset($data['ar_description'])) {
+            $data['description']['ar'] = $data['ar_description'];
+        }
+        if (isset($data['en_description'])) {
+            $data['description']['en'] = $data['en_description'];
+        }
 
         // Update the service data
         $service->update($data);
 
-        return new ServiceResource($service);
+        return $service;
     }
 
     /**
@@ -96,7 +132,12 @@ class ServService
      */
     public function delete($id, $force = null)
     {
+
         $service = Service::findOrFail($id);
+
+        if (auth()->user()->hasRole(Constants::SERVICE_PROVIDER_ROLE) && $service->service_provider_id !== auth()->id()) {
+            return error(__('messages.unauthorized_action'), [], 403);
+        }
 
         if ($force) {
             $service->forceDelete();
@@ -106,7 +147,6 @@ class ServService
 
         return response()->json([
             'success' => true,
-            'message' => 'Service deleted successfully.',
         ], 200);
     }
 
@@ -115,13 +155,15 @@ class ServService
      */
     public function restore(string|int $id)
     {
+       
         $service = Service::withTrashed()->find($id);
-
+        if (auth()->user()->hasRole(Constants::SERVICE_PROVIDER_ROLE) && $service->service_provider_id !== auth()->id()) {
+            return error(__('messages.unauthorized_action'), [], 403);
+        }
         if ($service && $service->trashed()) {
             $service->restore();
             return response()->json([
                 'success' => true,
-                'message' => 'Service restored successfully.',
             ], 200);
         }
 
@@ -134,38 +176,74 @@ class ServService
     public function show(int $id)
     {
         // Retrieve the service along with related category and provider data
-        $service = Service::with(['category', 'provider'])->findOrFail($id);
+        $service = Service::with(['category', 'provider', 'reviews'])->findOrFail($id);
+        if (auth()->user()->hasRole(Constants::SERVICE_PROVIDER_ROLE) && $service->service_provider_id !== auth()->id()) {
+            return error(__('messages.unauthorized_action'), [], 403);
+        }
 
-        return new ServiceResource($service);
+            // If the logged-in user is a normal user, return as ServiceResource
+        if (auth()->user()->hasRole(Constants::USER_ROLE)) {
+            return new ServiceResource($service);
+        }
+        return $service;
     }
 
     /**
-     * Get all services (used for listing).
+     * Change the status of a service.
      */
-    public function getAllServices()
+    public function changeStatus(ChangeServiceStatusRequest $request, int $id)
     {
-        $services = Service::with(['category', 'provider'])
-                           ->orderByDesc('created_at')
-                           ->get();
+        $service = Service::with(['category', 'provider', 'reviews'])->findOrFail($id);
 
-        return ServiceResource::collection($services);
+        if (auth()->user()->hasRole(Constants::SERVICE_PROVIDER_ROLE) && $service->service_provider_id !== auth()->id()) {
+            return error(__('messages.unauthorized_action'), [], 403);
+        }
+
+        $service->update(['status' => $request->status]);
+
+        return $service;
     }
 
-    /**
-     * Get the service categories (for filtering).
-     */
-    public function getCategories()
-    {
-        return Category::all(); // Return all categories to be used for filtering services
-    }
-    
-    /**
-     * Get manufacture years for services (if applicable).
-     */
-    public function getManufactureYears($serviceId)
-    {
-        $service = Service::findOrFail($serviceId);
+   
 
-        return $service->manufacture_years ?? [];
+
+        public function searchServices($searchTerm, $paginate = false, $limit = 10)
+    {
+        $query = Service::query()->with(['category', 'provider']);
+
+        $searchTerm = mb_strtolower($searchTerm, 'UTF-8');
+
+        $query->where(function ($q) use ($searchTerm) {
+            $q->whereRaw('LOWER(name->"$.ar") LIKE ?', ["%$searchTerm%"])
+            ->orWhereRaw('LOWER(name->"$.en") LIKE ?', ["%$searchTerm%"])
+            ->orWhereRaw('LOWER(description->"$.ar") LIKE ?', ["%$searchTerm%"])
+            ->orWhereRaw('LOWER(description->"$.en") LIKE ?', ["%$searchTerm%"]);
+        });
+
+        $query->orWhereHas('category', function ($q) use ($searchTerm) {
+            $q->whereRaw('LOWER(name) LIKE ?', ["%$searchTerm%"]);
+        });
+
+        $query->orWhereHas('provider', function ($q) use ($searchTerm) {
+            $q->whereRaw('LOWER(name) LIKE ?', ["%$searchTerm%"])
+            ->orWhere('phone', 'LIKE', "%$searchTerm%");
+        });
+
+        if (auth()->user()->hasRole(Constants::SERVICE_PROVIDER_ROLE)) {
+            $query->where('service_provider_id', auth()->id());
+        }
+
+        $query->orderByDesc('created_at');
+
+        if ($paginate) {
+            return $query->paginate($limit);
+        }
+
+        return $query->get();
     }
+
+
+
+        
+        
 }
